src/shipit_static_analysis/requirements_frozen.txt
aioamqp==0.10.0
aiohttp==2.3.3
appdirs==1.4.3
asn1crypto==0.23.0
async-timeout==1.4.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cookies==2.2.1
cryptography==2.1.3
elasticsearch==6.0.0
flake8==3.5.0
flake8-coding==1.3.0
flake8-quotes==0.12.0
google-api-python-client==1.6.4
httplib2==0.10.3
httpretty==0.8.14
icalendar==4.0.0
idna==2.6
libmozdata==0.1.39
Logbook==1.1.0
mccabe==0.6.1
mohawk==0.3.4
mozilla-cli-common==1.0.0
multidict==3.3.2
oauth2client==4.1.2
packaging==16.8
parsepatch==0.1.0
py==1.5.2
pyasn1==0.3.4
pyasn1-modules==0.1.5
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pyOpenSSL==17.4.0
pyparsing==2.2.0
pytest==3.2.5
python-dateutil==2.6.1
python-hglib==2.4
pytz==2017.3
PyYAML==3.12
raven==6.3.0
RBTools==0.7.10
requests==2.18.4
requests-futures==0.9.7
responses==0.8.1
rsa==3.4.2
six==1.10.0
slugid==1.0.7
structlog==17.2.0
taskcluster==2.0.0
tqdm==4.19.4
uritemplate==3.0.0
urllib3==1.22
whatthepatch==0.0.5
yarl==0.15.0
